requirements.txt
# This file was generated by generate_requirements.py
# Use 'pip install -r requirements.txt' to install all dependencies.

# --- Core Social Media API Interaction ---
praw==7.8.1
atproto==0.0.61

# --- LLM API Clients ---
openai==1.35.13
google-generativeai==0.7.2
anthropic==0.29.0

# --- Data Cleaning and Handling ---
beautifulsoup4==4.12.3
pandas==2.2.2
numpy==1.26.4

# --- Machine Learning & NLP (for Evaluation & Analysis) ---
scikit-learn==1.5.1
torch==2.4.0
transformers==4.42.4

# --- Utilities ---
python-dotenv==1.0.1

